requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
sqlalchemy==2.0.48
apscheduler==3.11.2
httpx==0.28.1
numpy==2.4.3
pandas==3.0.1
scikit-learn==1.8.0
torch==2.11.0
xgboost==3.2.0
lightgbm==4.6.0
ta==0.11.0
pytrends==4.9.2
yfinance==1.2.0
google-genai==1.69.0
